+++ src/UI/Data/Common/index.ts
@@ -1,1 +1,2 @@
+export * from "./AttributeConverterImpl";
 export * from "./AttributeConverter";
